cs1.3/requirements.txt
python-xlib>=0.29
colormath>=3.0.0
python-libxdo>=0.1.2a1
